warnings removed

+++ GraphicsProg2/Project/VulkanBase/Implementations/Initialization/DescriptorInit.cpp
@@ -52,7 +52,7 @@ void VulkanBase::CreateDescriptorPool()
 	poolInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
 	poolInfo.poolSizeCount = static_cast<uint32_t>(poolSizes.size());
 	poolInfo.pPoolSizes = poolSizes.data();
-	poolInfo.maxSets = m_Scene->GetMeshesAmount();
+	poolInfo.maxSets = static_cast<uint32_t>(m_Scene->GetMeshesAmount());
 
 	if (vkCreateDescriptorPool(m_Device, &poolInfo, nullptr, &m_DescriptorPool) != VK_SUCCESS)
 	{
